Lots of magic. TL;DR compiles on windows now

+++ src/r_segs.cpp
@@ -28,7 +28,7 @@ rcsid[] = "$Id: r_segs.c,v 1.3 1997/01/29 20:10:19 b1 Exp $";
 
 
 
-
+#include <cmath>
 #include <stdlib.h>
 
 #include "i_system.h"
@@ -399,7 +399,7 @@ R_StoreWallRange
     
     // calculate rw_distance for scale calculation
     rw_normalangle = curline->angle + ANG90;
-    offsetangle = abs(rw_normalangle-rw_angle1);
+    offsetangle = abs((int)(rw_normalangle-rw_angle1));
     
     if (offsetangle > ANG90)
 	offsetangle = ANG90;
